ensure dependees are computed
